+++ tests/unit/deploy_test.go
@@ -41,15 +41,12 @@ func TestDeployService_ErrorCreatingDirs(t *testing.T) {
 	assert.NotNil(t, err, "expected error calling Deploy()")
 }
 
-func TestDeployService_FileExists(t *testing.T) {
-
+func TestDeployService_DeclineToCreate(t *testing.T) {
 	newDir := "/newdir"
 	originDir := "/origindir"
 	filename := "/file.txt"
 	logfile := "/log.txt"
-	s3bucket := "mys3bucket"
-	lambdas := []string{"lambda1", "lambda2"}
-	codebuilds := []string{"codebuild1", "codebuild2"}
+	decline := "no"
 
 	emptyConfig := configs.Root{}
 	initMocks(emptyConfig)
@@ -59,6 +56,138 @@ func TestDeployService_FileExists(t *testing.T) {
 	mockFileSystemer.On("GetLocalMainTFFile").Return(filename)
 	mockFileSystemer.On("IsExistingFile", filename).Return(true)
 	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&decline)
+
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+
+	mockTerraformer.On("Init", mock.Anything, []string{}).Return(nil)
+
+	deployConfig := cfg.DeployConfig{
+		UseCached: true,
+	}
+	overrides := svc.DeployOverrides{
+		Namespace: "somethingpredictable",
+	}
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.Deploy(ctx, &overrides)
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+	mockAwser.AssertExpectations(t)
+	mockTFTemplater.AssertExpectations(t)
+	mockPrompter.AssertExpectations(t)
+
+	assert.NotNil(t, err, "expected use decline")
+}
+
+func TestDeployService_InitThrowsError(t *testing.T) {
+
+	newDir := "/newdir"
+	originDir := "/origindir"
+	filename := "/file.txt"
+	logfile := "/log.txt"
+	expectedErr := fmt.Errorf("error runninng init")
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+
+	mockFileSystemer.On("CreateConfigDirTree").Return(nil)
+	mockFileSystemer.On("ChToConfigDir").Return(newDir, originDir)
+	mockFileSystemer.On("GetLocalMainTFFile").Return(filename)
+	mockFileSystemer.On("IsExistingFile", filename).Return(true)
+	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+
+	mockTerraformer.On("Init", mock.Anything, []string{}).Return(expectedErr)
+
+	deployConfig := cfg.DeployConfig{
+		UseCached: true,
+	}
+	overrides := svc.DeployOverrides{
+		Namespace: "somethingpredictable",
+	}
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.Deploy(ctx, &overrides)
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+	mockAwser.AssertExpectations(t)
+	mockTFTemplater.AssertExpectations(t)
+
+	assert.NotNil(t, err, "expected error calling Deploy() when Init() errs")
+	assert.Equal(t, "error creating infrastructure: error runninng init", err.Error())
+}
+
+func TestDeployService_ApplyThrowsError(t *testing.T) {
+
+	newDir := "/newdir"
+	originDir := "/origindir"
+	filename := "/file.txt"
+	logfile := "/log.txt"
+	expectedErr := fmt.Errorf("error runninng apply")
+	affirm := "yes"
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+
+	mockFileSystemer.On("CreateConfigDirTree").Return(nil)
+	mockFileSystemer.On("ChToConfigDir").Return(newDir, originDir)
+	mockFileSystemer.On("GetLocalMainTFFile").Return(filename)
+	mockFileSystemer.On("IsExistingFile", filename).Return(true)
+	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&affirm)
+
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+
+	mockTerraformer.On("Init", mock.Anything, []string{}).Return(nil)
+	mockTerraformer.On("Apply", mock.Anything, []string{}).Return(expectedErr)
+
+	deployConfig := cfg.DeployConfig{
+		UseCached: true,
+	}
+	overrides := svc.DeployOverrides{
+		Namespace: "somethingpredictable",
+	}
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.Deploy(ctx, &overrides)
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+	mockAwser.AssertExpectations(t)
+	mockTFTemplater.AssertExpectations(t)
+	mockPrompter.AssertExpectations(t)
+
+	assert.NotNil(t, err, "expected error calling Deploy() when Apply() errors")
+	assert.Equal(t, "error creating infrastructure: error runninng apply", err.Error())
+}
+
+func TestDeployService_FileExists(t *testing.T) {
+
+	newDir := "/newdir"
+	originDir := "/origindir"
+	filename := "/file.txt"
+	logfile := "/log.txt"
+	s3bucket := "mys3bucket"
+	lambdas := []string{"lambda1", "lambda2"}
+	codebuilds := []string{"codebuild1", "codebuild2"}
+	affirm := "yes"
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+
+	mockFileSystemer.On("CreateConfigDirTree").Return(nil)
+	mockFileSystemer.On("ChToConfigDir").Return(newDir, originDir)
+	mockFileSystemer.On("GetLocalMainTFFile").Return(filename)
+	mockFileSystemer.On("IsExistingFile", filename).Return(true)
+	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&affirm)
 
 	mockFileSystemer.On("GetLogFile").Return(logfile)
 
@@ -89,6 +218,72 @@ func TestDeployService_FileExists(t *testing.T) {
 	mockTerraformer.AssertExpectations(t)
 	mockAwser.AssertExpectations(t)
 	mockTFTemplater.AssertExpectations(t)
+	mockPrompter.AssertExpectations(t)
+
+	assert.Nil(t, err, "expected no error calling Deploy() in happy path")
+}
+
+func TestDeployService_FileExistsWithOpts(t *testing.T) {
+
+	newDir := "/newdir"
+	originDir := "/origindir"
+	filename := "/file.txt"
+	logfile := "/log.txt"
+	s3bucket := "mys3bucket"
+	lambdas := []string{"lambda1", "lambda2"}
+	codebuilds := []string{"codebuild1", "codebuild2"}
+	affirm := "yes"
+	initOpts := []string{
+		"-backend-config=\"address=demo.consul.io\"",
+		"-backend-config=\"path=example_app/terraform_state\"",
+	}
+	applyOpts := []string{
+		"-compact-warnings",
+		"-backup=\"path\"",
+	}
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+
+	mockFileSystemer.On("CreateConfigDirTree").Return(nil)
+	mockFileSystemer.On("ChToConfigDir").Return(newDir, originDir)
+	mockFileSystemer.On("GetLocalMainTFFile").Return(filename)
+	mockFileSystemer.On("IsExistingFile", filename).Return(true)
+	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&affirm)
+
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+
+	mockTerraformer.On("Init", mock.Anything, initOpts).Return(nil)
+	mockTerraformer.On("Apply", mock.Anything, applyOpts).Return(nil)
+	mockTerraformer.On("GetOutput", mock.Anything, "artifacts_bucket_name").Return(s3bucket, nil)
+
+	mockFileSystemer.On("ChToTmpDir").Return(doesntMatter, doesntMatter)
+	mockGithuber.On("DownloadGithubReleaseAsset", "build_artifacts.zip")
+	mockFileSystemer.On("GetArtifactsDir").Return(doesntMatter)
+	mockFileSystemer.On("Unarchive", "build_artifacts.zip", doesntMatter)
+	mockFileSystemer.On("RemoveAll", "build_artifacts.zip")
+
+	mockAwser.On("UploadDirectoryToS3", doesntMatter, s3bucket, "").Return(lambdas, codebuilds)
+	mockAwser.On("UpdateLambdasFromS3Assets", lambdas, s3bucket, "somethingpredictable")
+
+	deployConfig := cfg.DeployConfig{
+		UseCached:      true,
+		TFInitOptions:  "-backend-config=\"address=demo.consul.io\" -backend-config=\"path=example_app/terraform_state\"",
+		TFApplyOptions: "-compact-warnings     -backup=\"path\"",
+	}
+	overrides := svc.DeployOverrides{
+		Namespace: "somethingpredictable",
+	}
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.Deploy(ctx, &overrides)
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+	mockAwser.AssertExpectations(t)
+	mockTFTemplater.AssertExpectations(t)
 
 	assert.Nil(t, err, "expected no error calling Deploy() in happy path")
 }
@@ -102,6 +297,7 @@ func TestDeployService_DoesNotFileExist(t *testing.T) {
 	s3bucket := "mys3bucket"
 	lambdas := []string{"lambda1", "lambda2"}
 	codebuilds := []string{"codebuild1", "codebuild2"}
+	affirm := "yes"
 
 	emptyConfig := configs.Root{}
 	initMocks(emptyConfig)
@@ -119,6 +315,8 @@ func TestDeployService_DoesNotFileExist(t *testing.T) {
 
 	// then everything resumes along the happy path as before...
 	mockFileSystemer.On("Chdir", originDir).Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&affirm)
 
 	mockFileSystemer.On("GetLogFile").Return(logfile)
 
@@ -147,6 +345,7 @@ func TestDeployService_DoesNotFileExist(t *testing.T) {
 	mockTerraformer.AssertExpectations(t)
 	mockAwser.AssertExpectations(t)
 	mockTFTemplater.AssertExpectations(t)
+	mockPrompter.AssertExpectations(t)
 
 	assert.Nil(t, err, "expected no error calling Deploy() in happy path")
 }
@@ -160,6 +359,7 @@ func TestDeployService_DoesNotFileExistAndUsingLocalRepo(t *testing.T) {
 	s3bucket := "mys3bucket"
 	lambdas := []string{"lambda1", "lambda2"}
 	codebuilds := []string{"codebuild1", "codebuild2"}
+	affirm := "yes"
 
 	emptyConfig := configs.Root{}
 	initMocks(emptyConfig)
@@ -172,6 +372,8 @@ func TestDeployService_DoesNotFileExistAndUsingLocalRepo(t *testing.T) {
 	// file is being created...
 	mockFileSystemer.On("ReadFromFile", mock.Anything).Return("filecontents")
 	mockFileSystemer.On("WriteFile", "/file.txt", "filecontents").Return()
+
+	mockPrompter.On("PromptBasic", mock.Anything, mock.Anything).Return(&affirm)
 
 	// then everything resumes along the happy path as before...
 	mockFileSystemer.On("Chdir", originDir).Return()
@@ -202,6 +404,102 @@ func TestDeployService_DoesNotFileExistAndUsingLocalRepo(t *testing.T) {
 	mockFileSystemer.AssertExpectations(t)
 	mockTerraformer.AssertExpectations(t)
 	mockAwser.AssertExpectations(t)
+	mockPrompter.AssertExpectations(t)
 
 	assert.Nil(t, err, "expected no error calling Deploy() in happy path")
+}
+
+func TestDeployService_PostDeployDefault(t *testing.T) {
+	logfile := "/log.txt"
+	apiURL := "https://some-api-id.execute-api.us-east-1.amazonaws.com/api"
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+	deployConfig := cfg.DeployConfig{
+		TFInitOptions:  "",
+		TFApplyOptions: "-compact-warnings",
+	}
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+	mockTerraformer.On("GetOutput", mock.Anything, "api_url").Return(apiURL, nil)
+
+	mockFileSystemer.On("WriteConfig").Return(nil)
+
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.PostDeploy(ctx)
+	assert.Nil(t, err)
+
+	assert.Equal(t, *service.Config.API.Host, "some-api-id.execute-api.us-east-1.amazonaws.com")
+	assert.Equal(t, *service.Config.API.BasePath, "/api")
+	// In default behavior, these should be left alone
+	assert.Nil(t, service.Config.Terraform.TFInitOptions, "TFInitOptions should be unset by default")
+	assert.Nil(t, service.Config.Terraform.TFApplyOptions, "TFApplyOptions should be unset by default")
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+}
+
+func TestDeployService_PostDeploySaveOpts(t *testing.T) {
+	logfile := "/log.txt"
+	apiURL := "https://some-api-id.execute-api.us-east-1.amazonaws.com/api"
+
+	emptyConfig := configs.Root{}
+	initMocks(emptyConfig)
+	deployConfig := cfg.DeployConfig{
+		SaveTFOptions:  true,
+		TFInitOptions:  "",
+		TFApplyOptions: "-compact-warnings",
+	}
+	mockFileSystemer.On("GetLogFile").Return(logfile)
+	mockTerraformer.On("GetOutput", mock.Anything, "api_url").Return(apiURL, nil)
+
+	mockFileSystemer.On("WriteConfig").Return(nil)
+
+	ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
+
+	err := service.PostDeploy(ctx)
+	assert.Nil(t, err)
+
+	assert.Equal(t, *service.Config.API.Host, "some-api-id.execute-api.us-east-1.amazonaws.com")
+	assert.Equal(t, *service.Config.API.BasePath, "/api")
+	// In default behavior, these should be left alone
+	assert.Equal(t, *service.Config.Terraform.TFInitOptions, deployConfig.TFInitOptions, "TFInitOptions should be set because save was specified")
+	assert.Equal(t, *service.Config.Terraform.TFApplyOptions, deployConfig.TFApplyOptions, "TFApplyOptions should be set because savw was specified")
+
+	mockFileSystemer.AssertExpectations(t)
+	mockTerraformer.AssertExpectations(t)
+}
+
+func Test_mockFileInfo_Name(t *testing.T) {
+	tests := []struct {
+		name string
+		m    *mockFileInfo
+		want string
+	}{
+		// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.m.Name(); got != tt.want {
+				t.Errorf("mockFileInfo.Name() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_mockFileInfo_IsDir(t *testing.T) {
+	tests := []struct {
+		name string
+		m    *mockFileInfo
+		want bool
+	}{
+		// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.m.IsDir(); got != tt.want {
+				t.Errorf("mockFileInfo.IsDir() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }