+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## vNext
+- Added `--tf-init-options` and `--tf-apply-options` for greater control over underlying provisioner
+- Added `--save-options` flag to persist the `--tf-init-options` and `--tf-apply-options` to the
+  configuration file when specified. Default behavior is false.
+- Persist API host name and path after running `system deploy`
+- **Potential breaking change**: explicitly configured logging output to go to STDOUT; command output should go to STDOUT
+
 ## v0.3.1
 - Moved quick start to readthedocs
 - Download deployment assets from the public url rather than using Github's GraphQL API